LayoutTests/http/tests/security/dataURL/xss-DENIED-from-javascript-url-window-open-expected.txt

CONSOLE MESSAGE: SecurityError: Blocked a frame with origin "null" from accessing a cross-origin frame. Protocols, domains, and ports must match.
Opener Frame

PASS: Access from a window opened with a data: URL was denied.
